pyproject.toml
[project]
name = "discourse-rss-share"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.28",
    "PyYAML>=6.0",
]

[project.scripts]
rss-share = "rss_share.main:main"
